CG-0MP00BL0Q003SLEO: Decompose SushiGoScene into helper classes
Extract responsibilities from 1423-line SushiGoScene.ts into composable helpers:
- SushiGoConstants.ts: shared layout, styling, and audio constants
- SushiGoRenderer.ts: UI creation and basic refresh logic
- SushiGoCardFactory.ts: visual card rectangle creation
- SushiGoTableauRenderer.ts: tableau grouping and rendering
- SushiGoOverlayManager.ts: round score and game over overlays
- SushiGoTooltipManager.ts: card tooltip display
- SushiGoReplayController.ts: replay mode state injection

SushiGoScene.ts reduced from 1423 lines to 608 lines.
All tests pass (npm test and npm run build verified).
